+++ docs/changelog.rst
@@ -4,6 +4,11 @@ Changelog
 :version:`Unreleased <v3.36.1...HEAD>` - TBD
 --------------------------------------------
 
+**Changed**
+
+- Merge ``minLength`` & ``maxLength`` into ``pattern`` to avoid extremely slow generation in most popular cases.
+- Avoid generating ``{`` and ``}`` for path parameters.
+
 .. _v3.36.1:
 
 :version:`3.36.1 <v3.36.0...v3.36.1>` - 2024-09-23